WiP: go.{mod,sum}: temporarily divert NRI and re-vendor.
Temporarily divert NRI and re-vendor. This commit diverts
NRI to the latest main/HEAD at the time of this commit. To
test with NRI features unavailable in the main repo, update
the replacement rule to point to a repo with the changes.
You can then test your changes with 'make vendor && make'.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>